Client/requirements.txt
grpcio
pandas
flask
redis
googleapis-common-protos